requirements.txt
Babel==2.12.1
Flask==2.2.3
Flask_Login==0.6.2
Flask_WTF==1.1.1
mysql_connector_repackaged==0.3.1
pyecharts==2.0.2
Werkzeug==2.2.3
WTForms==2.3.3
